// src/store/item/item.types.ts
export enum ITEM_ACTION_TYPES {
  // ADD ITEM
  ADD_ITEM_START = "item/ADD_ITEM_START",
  ADD_ITEM_SUCCESS = "item/ADD_ITEM_SUCCESS",
  ADD_ITEM_FAILED = "item/ADD_ITEM_FAILED",

  // GET ITEMS
  FETCH_ITEMS_START = "item/FETCH_ITEMS_START",
  FETCH_ITEMS_SUCCESS = "item/FETCH_ITEMS_SUCCESS",
  FETCH_ITEMS_FAILED = "item/FETCH_ITEMS_FAILED",

  // UPDATE ITEM
  UPDATE_ITEM_START = "item/UPDATE_ITEM_START",
  UPDATE_ITEM_SUCCESS = "item/UPDATE_ITEM_SUCCESS",
  UPDATE_ITEM_FAILED = "item/UPDATE_ITEM_FAILED",

  // DELETE ITEM
  DELETE_ITEM_START = "item/DELETE_ITEM_START",
  DELETE_ITEM_SUCCESS = "item/DELETE_ITEM_SUCCESS",
  DELETE_ITEM_FAILED = "item/DELETE_ITEM_FAILED",

  // LAST N DAYS ITEMS
  FETCH_LAST_N_DAYS_ITEMS_START = "item/FETCH_LAST_N_DAYS_ITEMS_START",
  FETCH_LAST_N_DAYS_ITEMS_SUCCESS = "item/FETCH_LAST_N_DAYS_ITEMS_SUCCESS",
  FETCH_LAST_N_DAYS_ITEMS_FAILED = "item/FETCH_LAST_N_DAYS_ITEMS_FAILED",
}